requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
requests==2.31.0
ftfy==6.1.3
cloudscraper==1.2.71
lxml==5.1.0
